+++ src/main/java/com/vizcainoraw/hotelalura/model/Reservation.java
@@ -14,7 +14,7 @@ public class Reservation {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer ReservationId;
+    private Long ReservationId;
 
     @ManyToOne
     @JoinColumn(name = "customer_id", nullable = false)
