+++ src/main/java/com/vizcainoraw/hotelalura/repository/ServiceRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.vizcainoraw.hotelalura.model.Service;
 
 @Repository
-public interface ServiceRepository extends JpaRepository<Service, Integer>{
+public interface ServiceRepository extends JpaRepository<Service, Long>{
     
 }
